==> 1-data-analysis-and-experimentation.ipynb <==
import pandas as pd

### Import Data

df_train = pd.read_csv('training.txt', header=None, names=['user_id', 'item_id', 'rating'])
df_test = pd.read_csv('testing.txt', header=None, names=['user_id', 'item_id', 'rating'])

with open('item_tag.txt', 'r') as f:
    lines = f.readlines()

df_tag = pd.DataFrame({'item_id': [int(line.split(':')[0]) for line in lines],
                   'categories': [[int(x) for x in line.split(':')[1].split(',')] for line in lines]})


df_tag_dupes = df_tag['item_id'].duplicated()
df_tag_dupes.sum() # no duplicates of ids in the df_tag data

df_train = df_train.sort_values(by='user_id')
df_test = df_test.sort_values(by='user_id')
df_tag = df_tag.sort_values(by='item_id')

### Checking if users in test are not in train

train_user_id = df_train['user_id'].drop_duplicates()
test_user_id = df_test['user_id'].drop_duplicates()

train_user_id.shape

check = test_user_id.isin(train_user_id)

test_user_id.shape[0] == test_user_id[check].shape[0] #in further parts I also found that there are items in the test set that aren't present in the train set

all user_ids in test set are in training set

### Checking ratings

df_train['rating'].drop_duplicates().sort_values()

from surprise import Reader, Dataset

reader = Reader(rating_scale=(1, 5))
data_train = Dataset.load_from_df(df_train, reader)

### Testing base methods

from surprise import BaselineOnly, SVD, SVDpp, NMF
from surprise.model_selection import cross_validate

# baseline method
baseline = BaselineOnly()
cross_validate(baseline, data_train, measures=["RMSE"], cv=3, verbose=True)

the baseline methods takes a long time just to train even using cloud computing thus it is unwise to use the whole dataset for experimentation. I will select a subset of the dataset that is reprentative of the population

The steps to ensure this is:
I use Random User Sampling. This technique is based on the paper On Sampling Collaborative Filtering Datasets by Noveen Sachdeva et al. The researches state

" In Random User Sampling, we retain users from D at random.
To be more specific, we iteratively preserve all the interactions
for a random user until we have retained 𝑝% of the original interactions."

D is the 
user_id, item_id, rating table

### Chosing sample of data for experimentation

import numpy as np

train_user_id_set = set(train_user_id.copy().reset_index(drop=True))

user_to_review_count = df_train.groupby('user_id').size()

review_count = df_train.shape[0]

user_to_review_count_hash = {}

for uid, val in enumerate(user_to_review_count):
    user_to_review_count_hash[uid] = val

import random

percentage_retained = 0.30 #choosing p as 30%
current_review_count = 0
chosen_user_ids = []
while current_review_count < review_count * percentage_retained:
    chosen_user_id = random.sample(train_user_id_set,1)[0]
    train_user_id_set.remove(chosen_user_id)
    chosen_user_ids.append(chosen_user_id)
    current_review_count += user_to_review_count_hash[chosen_user_id]

df_train_sample = df_train[df_train['user_id'].isin(set(chosen_user_ids))]

current_review_count / review_count

current_review_count

df_train_sample.to_csv('sample_train.csv')

### Testing base methods on sample

df_train_sample = pd.read_csv('sample_train.csv')

df_train_sample = df_train_sample[['user_id','item_id','rating']]

data_train_sample = Dataset.load_from_df(df_train_sample, reader)

import time

# baseline method
start_time = time.time()
baseline = BaselineOnly()
cross_validate(baseline, data_train_sample, measures=["RMSE"], cv=3, verbose=True)
end_time = time.time()

time.strftime('%H:%M:%S', time.gmtime(end_time - start_time))

# SVD Method
start_time_svd = time.time()
svd = SVD()
cross_validate(svd, data_train_sample, measures=["RMSE"], cv=3, verbose=True)
end_time_svd = time.time()

time.strftime('%H:%M:%S', time.gmtime(end_time_svd - start_time_svd))

# SVDpp Method
start_time_svdpp = time.time()
svdpp = SVDpp()
cross_validate(svdpp, data_train_sample, measures=["RMSE"], cv=3, verbose=True)
end_time_svdpp = time.time()

time.strftime('%H:%M:%S', time.gmtime(end_time_svdpp - start_time_svdpp))

# nmf Method
start_time_nmf = time.time()
nmf = NMF()
cross_validate(nmf, data_train_sample, measures=["RMSE"], cv=3, verbose=True)
end_time_nmf = time.time()

time.strftime('%H:%M:%S', time.gmtime(end_time_nmf - start_time_nmf))

best RMSE in order
- SVD++
- SVD
- Baseline
- NMF

to proceed with hyperparameter tuning I chose SVD instead of the SVD++ because of SVD++'s high train time
