==> 3-hypertuning-analysis.ipynb <==
import pandas as pd
import numpy as np
import ast

df_tuning = pd.read_csv('81.csv')

df_tuning = df_tuning.iloc[:,1:]

df_tuning['result'] = df_tuning['result'].apply(lambda x: ast.literal_eval(x.replace("\'", "\"").replace('array', '')))

df_tuning['mean'] = df_tuning['result'].apply(lambda x: np.mean(x['test_rmse']))

df_tuning.sort_values(by='mean',inplace=True)

df_tuning.head()

#lets train on whole dataset
