==> 4-svd-training.ipynb <==
import pandas as pd

### Import Data

df_train = pd.read_csv('training.txt', header=None, names=['user_id', 'item_id', 'rating'])
df_test = pd.read_csv('testing.txt', header=None, names=['user_id', 'item_id', 'rating'])

df_train = df_train.sort_values(by='user_id')
df_test = df_test.sort_values(by='user_id')

### Checking if users in test are not in train

train_user_id = df_train['user_id'].drop_duplicates()
test_user_id = df_test['user_id'].drop_duplicates()

train_user_id.shape

check = test_user_id.isin(train_user_id)

test_user_id.shape[0] == test_user_id[check].shape[0]

all user_ids in test set are in training set

### Checking ratings

df_train['rating'].drop_duplicates().sort_values()

from surprise import Reader, Dataset, accuracy, Reader
from surprise import SVD
from surprise.model_selection import cross_validate

reader = Reader(rating_scale=(1, 5))
data_train = Dataset.load_from_df(df_train, reader)

### Testing methods on MF

# SVD Method
print('running svd')
svd = SVD()
cross_validate(svd, data_train, measures=["RMSE"], cv=3, verbose=True)

### 10 Fold Crossvalidation

import time

# SVD Tuning
start_time_svd_tune = time.time()

svd = SVD(n_factors=80, n_epochs = 60, lr_all=0.005, reg_all=0.05)
result = cross_validate(svd, data_train, measures=["RMSE"], cv=10, verbose=True)


end_time_svd_tune = time.time()
time.strftime('%H:%M:%S', time.gmtime(end_time_svd_tune - start_time_svd_tune))

mean of 10-fold cross validation = 0.7779

data_train_fit = data_train.build_full_trainset()

svd = SVD(n_factors=80, n_epochs = 60, lr_all=0.005, reg_all=0.05)
svd.fit(data_train_fit)

from surprise.dump import dump, load

dump('svd_final.pkl',None,svd,1)
