==> 5-prediction.ipynb <==
from surprise.dump import dump, load
import pandas as pd

svd = load('svd_final.pkl')

svd_model = svd[1]

df_train = pd.read_csv('training.txt', header=None, names=['user_id', 'item_id', 'rating'])
df_test = pd.read_csv('testing.txt', header=None, names=['user_id', 'item_id', 'rating'])

predictions = []
for index, row in df_test.iterrows():
    user = row['user_id']
    item = row['item_id']
    predictions.append(svd_model.predict(user,item)[3])

df_test['rating'] = predictions

df_test.to_csv('prediction.txt', index=False, header=False)
